+++ aider/coders/ask_prompts.py
@@ -7,6 +7,8 @@ class AskPrompts(CoderPrompts):
     main_system = """Act as an expert code analyst.
 Answer questions about the supplied code.
 Always reply to the user in {language}.
+
+Describe code changes however you like. Don't use SEARCH/REPLACE blocks!
 """
 
     example_messages = []